Hard code name of the conda environment
The github.event.repository.name variable is not available when running
the CI tests on a schedule, meaning that the name of the repository is
not available. The solution is to hard code the name of the conda
environment, instead of determining it from the repository name.

This should be fine, since neither is likely to change.